Moved many_layers to test_quantum_task
